// src/lib/generated/Rule.ts
// This file was generated by [ts-rs](https://github.com/Aleph-Alpha/ts-rs). Do not edit this file manually.
import type { Action } from "./Action";
import type { Predicate } from "./Predicate";

/**
 * A stored rule.
 */
export type Rule = { id: number, name: string, enabled: boolean, predicate: Predicate, actions: Array<Action>, sortOrder: number, };